+++ R/globals.R
@@ -1,0 +1,1 @@
+globalVariables(c("name", "path", "size", "modification_time", "birth_time", "active", "changed", "expires", "path"))